Move Supabase path to ./lib
