+++ src/app/api/auth/login/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
-import { supabaseServer } from "@/src/lib/api/supabase";
+import { supabaseServer } from "@/src/lib/supabase";
 
 export async function POST(req: NextRequest) {
   try {
